+++ flake.nix
@@ -7,12 +7,15 @@
   inputs = {
     nixpkgs.url = "github:NixOS/nixpkgs?ref=nixos-23.05";
     pre-commit-hooks.url = "github:cachix/pre-commit-hooks.nix";
+    horizon-core.url = "git+https://gitlab.horizon-haskell.net/package-sets/horizon-core";
     autodocodec.url = "github:NorfairKing/autodocodec";
     autodocodec.flake = false;
-    validity.url = "github:NorfairKing/validity";
+    validity.url = "github:NorfairKing/validity?ref=forward-compatibility";
     validity.flake = false;
     safe-coloured-text.url = "github:NorfairKing/safe-coloured-text";
     safe-coloured-text.flake = false;
+    fast-myers-diff.url = "github:NorfairKing/fast-myers-diff";
+    fast-myers-diff.flake = false;
     nixpkgs-22_11.url = "github:NixOS/nixpkgs?ref=nixos-22.11";
     nixpkgs-22_05.url = "github:NixOS/nixpkgs?ref=nixos-22.05";
     nixpkgs-21_11.url = "github:NixOS/nixpkgs?ref=nixos-21.11";
@@ -25,27 +28,44 @@
     , nixpkgs-22_05
     , nixpkgs-21_11
     , pre-commit-hooks
+    , horizon-core
     , autodocodec
     , validity
     , safe-coloured-text
+    , fast-myers-diff
     }:
     let
       system = "x86_64-linux";
+      overlays = [
+        self.overlays.${system}
+        (import (validity + "/nix/overlay.nix"))
+        (import (autodocodec + "/nix/overlay.nix"))
+        (import (safe-coloured-text + "/nix/overlay.nix"))
+        (import (fast-myers-diff + "/nix/overlay.nix"))
+      ];
       pkgsFor = nixpkgs: import nixpkgs {
         inherit system;
         config.allowUnfree = true;
+        inherit overlays;
+      };
+
+      horizonPkgs = import nixpkgs {
+        inherit system;
         overlays = [
-          self.overlays.${system}
-          (import (validity + "/nix/overlay.nix"))
-          (import (autodocodec + "/nix/overlay.nix"))
-          (import (safe-coloured-text + "/nix/overlay.nix"))
-        ];
+          (final: prev: {
+            haskellPackages = prev.haskellPackages.override (old: {
+              overrides = final.lib.composeExtensions (old.overrides or (_: _: { })) (self: super:
+                horizon-core.legacyPackages.${system} // super
+              );
+            });
+          })
+        ] ++ overlays;
       };
       pkgs = pkgsFor nixpkgs;
     in
     {
       overlays.${system} = import ./nix/overlay.nix;
-      packages.${system}.default = pkgs.haskellPackages.sydtestRelease;
+      packages.${system} = pkgs.haskellPackages.sydtestPackages;
       checks.${system} =
         let
           backwardCompatibilityCheckFor = nixpkgs:
@@ -60,7 +80,8 @@
           backwardCompatibilityChecks = pkgs.lib.mapAttrs (_: nixpkgs: backwardCompatibilityCheckFor nixpkgs) allNixpkgs;
         in
         backwardCompatibilityChecks // {
-          release = self.packages.${system}.default;
+          forwardCompatibility = horizonPkgs.haskellPackages.sydtestRelease;
+          release = pkgs.haskellPackages.sydtestRelease;
           shell = self.devShells.${system}.default;
           pre-commit = pre-commit-hooks.lib.${system}.run {
             src = ./.;