+++ sydtest-discover/default.nix
@@ -3,7 +3,7 @@
 }:
 mkDerivation {
   pname = "sydtest-discover";
-  version = "0.0.0.3";
+  version = "0.0.0.4";
   src = ./.;
   isLibrary = true;
   isExecutable = true;